feat: add rendering thread

+++ RayTracing/include/utils.h
@@ -11,7 +11,7 @@
 using std::make_shared;
 using std::shared_ptr;
 
-using real = double;
+using real = float;
 
 constexpr double infinity   = std::numeric_limits<double>::infinity();
 constexpr double pi         = 3.1415926535897932385;
